# server/oapi/types.go
// Package oapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.13.4 DO NOT EDIT.
package oapi

import (
	"time"

	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
)

const (
	BasicAuthScopes = "basicAuth.Scopes"
)

// Defines values for UserRole.
const (
	Editor UserRole = "editor"
	Viewer UserRole = "viewer"
)

// Todo defines model for Todo.
type Todo struct {
	CreatedAt   time.Time           `json:"created_at"`
	CreatedBy   *int                `json:"created_by,omitempty"`
	Description *string             `json:"description,omitempty"`
	DueDate     *openapi_types.Date `json:"due_date,omitempty"`
	Id          int                 `json:"id"`
	Title       string              `json:"title"`
	UpdatedAt   time.Time           `json:"updated_at"`
}

// User defines model for User.
type User struct {
	CreatedAt time.Time `json:"created_at"`
	Id        int       `json:"id"`
	Name      string    `json:"name"`
	Role      *UserRole `json:"role,omitempty"`
	Todos     Todo      `json:"todos"`
	UpdatedAt time.Time `json:"updated_at"`
}

// UserRole defines model for User.Role.
type UserRole string

// TodoId defines model for todo_id.
type TodoId = int

// UserId defines model for user_id.
type UserId = int

// CreateTodoJSONBody defines parameters for CreateTodo.
type CreateTodoJSONBody struct {
	Description *string             `json:"description,omitempty"`
	DueDate     *openapi_types.Date `json:"due_date,omitempty"`
	Title       string              `json:"title"`
}

// UpdateTodoJSONBody defines parameters for UpdateTodo.
type UpdateTodoJSONBody struct {
	Description *string             `json:"description,omitempty"`
	DueDate     *openapi_types.Date `json:"due_date,omitempty"`
	Title       string              `json:"title"`
}

// CreateUserJSONBody defines parameters for CreateUser.
type CreateUserJSONBody struct {
	Name string `json:"name"`
}

// UpdateUserJSONBody defines parameters for UpdateUser.
type UpdateUserJSONBody struct {
	Name string `json:"name"`
}

// CreateTodoJSONRequestBody defines body for CreateTodo for application/json ContentType.
type CreateTodoJSONRequestBody CreateTodoJSONBody

// UpdateTodoJSONRequestBody defines body for UpdateTodo for application/json ContentType.
type UpdateTodoJSONRequestBody UpdateTodoJSONBody

// CreateUserJSONRequestBody defines body for CreateUser for application/json ContentType.
type CreateUserJSONRequestBody CreateUserJSONBody

// UpdateUserJSONRequestBody defines body for UpdateUser for application/json ContentType.
type UpdateUserJSONRequestBody UpdateUserJSONBody